tests/turbohtml-tests/plaintext_element.dat

#data
<plaintext>test<b>
#errors
#document
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "test<b>"
